refactor: mv basis path in module
