+++ qlinks/symmetry/gauss_law.py
@@ -18,7 +18,7 @@ from qlinks.exceptions import (
 from qlinks.lattice.component import Site, UnitVectors
 from qlinks.lattice.square_lattice import SquareLattice, Vertex
 from qlinks.solver.deep_first_search import Node
-from qlinks.symmetry.computation_basis import ComputationBasis
+from qlinks.computation_basis import ComputationBasis
 
 
 class Flow(IntEnum):
