+++ qlinks/symmetry/translation.py
@@ -13,7 +13,7 @@ import scipy.sparse as sp
 from qlinks.exceptions import InvalidArgumentError
 from qlinks.lattice.component import UnitVector
 from qlinks.lattice.square_lattice import LocalOperator, SquareLattice
-from qlinks.symmetry.computation_basis import ComputationBasis
+from qlinks.computation_basis import ComputationBasis
 
 
 @dataclass(slots=True)
