+++ tests/symmetry/test_translation.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from qlinks.lattice.square_lattice import SquareLattice
-from qlinks.symmetry.computation_basis import ComputationBasis
+from qlinks.computation_basis import ComputationBasis
 from qlinks.symmetry.translation import Translation
